Fixing panics in test262 when running in debug mode (#157)
This PR fixes nearly all of the panics that occur when running
temporal_rs through the test262 suite using Boa as noted by @jedel1043.

+++ src/options/relative_to.rs
@@ -68,7 +68,7 @@ impl RelativeTo {
             }
             TimeZoneRecord::Offset(offset_record) => {
                 // NOTE: ixdtf parser restricts minute/second to 0..=60
-                let minutes = i16::from((offset_record.hour * 60) + offset_record.minute);
+                let minutes = i16::from(offset_record.hour) * 60 + offset_record.minute as i16;
                 TimeZone::OffsetMinutes(minutes * i16::from(offset_record.sign as i8))
             }
             // TimeZoneRecord is non_exhaustive, but all current branches are matching.
